server/requirements-cpu.txt
connexion[swagger-ui]
keras
tensorflow
Pillow
numpy
h5py
tornado